+++ test/test.js
@@ -1,9 +1,9 @@
 require ('./components/calculateFilePathTest.js');
 require ('./components/extractFieldsTest.js');
-require ('./components/filterOutDeprecatedRecordsTest.js');
 require ('./components/filterOutIncompleteRecordsTest.js');
 require ('./components/filterOutNamelessRecordsTest.js');
 require ('./components/filterOutUnreadableFilesTest.js');
+require ('./components/isNotDeprecatedRecordTest.js');
 require ('./components/isValidIdTest.js');
 require ('./components/loadJSONTest.js');
 require ('./hierarchyFinderTest.js');
